docs: change frequency response graph from InteractiveExample to AudioApiExample
