+++ packages/audiodocs/src/examples/Biquad/FrequencyResponseGraph.tsx
@@ -14,44 +14,14 @@ import {
   ActivityIndicator,
   ScrollView,
 } from 'react-native';
+import { drawFrequencyResponse } from './drawFrequencyResponse';
+import RangeSlider from './RangeSlider';
 
 const FILTER_TYPES: BiquadFilterType[] = [
   'allpass', 'lowpass', 'highpass', 'bandpass',
   'lowshelf', 'highshelf', 'peaking', 'notch',
 ];
 
-const RangeSlider = ({
-  label,
-  value,
-  min,
-  max,
-  step,
-  onChange,
-  unit = '',
-}: {
-  label: string;
-  value: number;
-  min: number;
-  max: number;
-  step: number;
-  unit?: string;
-  onChange: (v: number) => void;
-}) => (
-  <div style={{ marginTop: 8 }}>
-    <label style={{ color: '#fcfcff' }}>
-      {label}: {value.toFixed(1)} {unit}
-      <input
-        type="range"
-        min={min}
-        max={max}
-        step={step}
-        value={value}
-        onChange={(e) => onChange(Number(e.target.value))}
-        style={{ width: '100%', marginTop: 4, color: '#6676aa' }}
-      />
-    </label>
-  </div>
-);
 
 const FrequencyResponseGraph: React.FC = () => {
   const [isPlaying, setIsPlaying] = useState(false);
@@ -86,7 +56,7 @@ const FrequencyResponseGraph: React.FC = () => {
           .then((response) => response.arrayBuffer())
           .then((arrayBuffer) => ctx.decodeAudioData(arrayBuffer));
 
-        const audio2 = await fetch('/react-native-audio-api/audio/music/example-music-01.mp3')
+        const audio2 = await fetch('/react-native-audio-api/audio/music/example-music-03.mp3')
           .then((response) => response.arrayBuffer())
           .then((arrayBuffer) => ctx.decodeAudioData(arrayBuffer));
 
@@ -112,7 +82,7 @@ const FrequencyResponseGraph: React.FC = () => {
     filter.Q.value = filterQ;
     filter.gain.value = filterGain;
 
-    drawFrequencyResponse();
+    drawFrequencyResponse(canvasRef.current, filterRef.current);
   }, [filterType, filterFreq, filterQ, filterGain]);
 
   const playAudio = async () => {
@@ -147,60 +117,6 @@ const FrequencyResponseGraph: React.FC = () => {
   useEffect(() => {
     if (isPlaying) playAudio();
   }, [selectedAudio]);
-
-  const drawFrequencyResponse = () => {
-    if (!canvasRef.current || !filterRef.current) return;
-
-    const ctx = canvasRef.current.getContext('2d');
-    if (!ctx) return;
-
-    const { width, height } = canvasRef.current;
-
-    const N = 200;
-    const frequencies = new Float32Array(N).map((_, i) => 20 * 1000 ** (i / (N - 1)));
-    const mags = new Float32Array(N);
-    const phases = new Float32Array(N);
-    filterRef.current.getFrequencyResponse(frequencies, mags, phases);
-
-    // grid
-    const line = (x1: number, y1: number, x2: number, y2: number) => {
-      ctx.beginPath();
-      ctx.moveTo(x1, y1);
-      ctx.lineTo(x2, y2);
-      ctx.stroke();
-    };
-
-    ctx.clearRect(0, 0, width, height);
-    ctx.strokeStyle = '#bcdae4ff';
-    ctx.lineWidth = 1;
-
-    const numGridLines = 11;
-    for (let i = 0; i < numGridLines; i++) {
-      const x = (i / (numGridLines - 1)) * width;
-      line(x, 0, x, height);
-    }
-
-    const dBMarks = [-30, -20, -10, 0, 10, 20, 30];
-    for (let i = 0; i < dBMarks.length; i++) {
-      const y = (i / (dBMarks.length - 1)) * height;
-      line(0, y, width, y);
-    }
-
-    // frequency response
-    ctx.beginPath();
-    mags.forEach((m, i) => {
-      // Map the index (0..N) to the horizontal axis
-      const x = (i / (mags.length - 1)) * width;
-      const db = 20 * Math.log10(m);
-      // Map dB range [-30, +30] to vertical axis
-      const y = height - ((db + 30) / 60) * height;
-      i === 0 ? ctx.moveTo(x, y) : ctx.lineTo(x, y);
-    });
-
-    ctx.strokeStyle = '#38acdd';
-    ctx.lineWidth = 2;
-    ctx.stroke();
-  };
 
   return (
     <View style={styles.container}>
