--- packages/audiodocs/src/examples/Biquad/FrequencyResponseGraphSource.tsx
@@ -1,299 +1,0 @@
-import React, { useState, useEffect, useRef } from 'react';
-import {
-  AudioContext,
-  AudioBuffer,
-  AudioBufferSourceNode,
-  BiquadFilterNode,
-} from 'react-native-audio-api';
-import {
-  View,
-  Button,
-  Text,
-  Pressable,
-  StyleSheet,
-  ActivityIndicator,
-  ScrollView,
-} from 'react-native';
-import Slider from '@react-native-community/slider';
-import Canvas from 'react-native-canvas';
-
-const FILTER_TYPES: BiquadFilterType[] = [
-  'allpass', 'lowpass', 'highpass', 'bandpass',
-  'lowshelf', 'highshelf', 'peaking', 'notch',
-];
-
-const RangeSlider = ({
-  label,
-  value,
-  min,
-  max,
-  step,
-  onChange,
-  unit = '',
-}: {
-  label: string;
-  value: number;
-  min: number;
-  max: number;
-  step: number;
-  unit?: string;
-  onChange: (v: number) => void;
-}) => (
-  <View style={{ marginTop: 8 }}>
-    <Text style={{ color: '#fcfcff' }}>
-      {label}: {value.toFixed(1)} {unit}
-    </Text>
-    <Slider
-      minimumValue={min}
-      maximumValue={max}
-      step={step}
-      value={value}
-      onValueChange={onChange}
-      minimumTrackTintColor="#38acdd"
-      maximumTrackTintColor="#6676aa"
-      thumbTintColor="#fcfcff"
-      style={{ width: '100%', marginTop: 4 }}
-    />
-  </View>
-);
-
-const FrequencyResponseGraph: React.FC = () => {
-  const [isPlaying, setIsPlaying] = useState(false);
-  const [isLoading, setIsLoading] = useState(false);
-  const [filterType, setFilterType] = useState<BiquadFilterType>('allpass');
-  const [filterFreq, setFilterFreq] = useState(350);
-  const [filterQ, setFilterQ] = useState(1);
-  const [filterGain, setFilterGain] = useState(0);
-  const [selectedAudio, setSelectedAudio] = useState<'audio1' | 'audio2'>('audio1');
-
-  const audioContextRef = useRef<AudioContext | null>(null);
-  const bufferSourceRef = useRef<AudioBufferSourceNode | null>(null);
-  const audioBuffersRef = useRef<{ audio1: AudioBuffer | null; audio2: AudioBuffer | null }>({
-    audio1: null,
-    audio2: null,
-  });
-  const filterRef = useRef<BiquadFilterNode | null>(null);
-  const canvasRef = useRef<Canvas | null>(null);
-
-  useEffect(() => {
-    const init = async () => {
-      const ctx = new AudioContext();
-      audioContextRef.current = ctx;
-
-      const filter = ctx.createBiquadFilter();
-      filter.connect(ctx.destination);
-      filterRef.current = filter;
-
-      setIsLoading(true);
-      try {
-        const audio1 = await fetch('https://software-mansion.github.io/react-native-audio-api/audio/voice/example-voice-01.mp3')
-          .then((response) => response.arrayBuffer())
-          .then((arrayBuffer) => ctx.decodeAudioData(arrayBuffer));
-
-        const audio2 = await fetch('https://software-mansion.github.io/react-native-audio-api/audio/music/example-music-01.mp3')
-          .then((response) => response.arrayBuffer())
-          .then((arrayBuffer) => ctx.decodeAudioData(arrayBuffer));
-
-        audioBuffersRef.current = { audio1, audio2 };
-      } catch (error) {
-        console.error('Error decoding audio data source:', error);
-      }
-      setIsLoading(false);
-    };
-
-    init();
-    return () => {
-      audioContextRef.current?.close();
-    };
-  }, []);
-
-  useEffect(() => {
-    if (!filterRef.current) return;
-
-    const filter = filterRef.current;
-    filter.type = filterType;
-    filter.frequency.value = filterFreq;
-    filter.Q.value = filterQ;
-    filter.gain.value = filterGain;
-
-    drawFrequencyResponse();
-  }, [filterType, filterFreq, filterQ, filterGain]);
-
-  const playAudio = async () => {
-    if (!audioContextRef.current) return;
-
-    bufferSourceRef.current?.stop();
-    bufferSourceRef.current = null;
-
-    const buffer = selectedAudio === 'audio1' ? audioBuffersRef.current.audio1 : audioBuffersRef.current.audio2;
-    if (!buffer) return;
-
-    const source = await audioContextRef.current.createBufferSource();
-    source.buffer = buffer;
-    bufferSourceRef.current = source;
-
-    source.connect(filterRef.current!);
-    source.start();
-
-    setIsPlaying(true);
-  };
-
-  const handlePlayPause = async () => {
-    if (isPlaying) {
-      bufferSourceRef.current?.stop();
-      bufferSourceRef.current = null;
-      setIsPlaying(false);
-    } else {
-      await playAudio();
-    }
-  };
-
-  useEffect(() => {
-    if (isPlaying) playAudio();
-  }, [selectedAudio]);
-
-  const drawFrequencyResponse = () => {
-    if (!canvasRef.current || !filterRef.current) return;
-
-    const canvas = canvasRef.current;
-    const ctx = canvas.getContext('2d');
-    if (!ctx) return;
-
-    const width = 600;
-    const height = 200;
-
-    const N = 200;
-    const frequencies = new Float32Array(N).map((_, i) => 20 * 1000 ** (i / (N - 1)));
-    const mags = new Float32Array(N);
-    const phases = new Float32Array(N);
-    filterRef.current.getFrequencyResponse(frequencies, mags, phases);
-
-    // grid
-    const line = (x1: number, y1: number, x2: number, y2: number) => {
-      ctx.beginPath();
-      ctx.moveTo(x1, y1);
-      ctx.lineTo(x2, y2);
-      ctx.stroke();
-    };
-
-    ctx.clearRect(0, 0, width, height);
-    ctx.strokeStyle = '#bcdae4ff';
-    ctx.lineWidth = 1;
-
-    const numGridLines = 11;
-    for (let i = 0; i < numGridLines; i++) {
-      const x = (i / (numGridLines - 1)) * width;
-      line(x, 0, x, height);
-    }
-
-    const dBMarks = [-30, -20, -10, 0, 10, 20, 30];
-    for (let i = 0; i < dBMarks.length; i++) {
-      const y = (i / (dBMarks.length - 1)) * height;
-      line(0, y, width, y);
-    }
-
-    // frequency response
-    ctx.beginPath();
-    mags.forEach((m, i) => {
-      // Map the index (0..N) to the horizontal axis
-      const x = (i / (mags.length - 1)) * width;
-      const db = 20 * Math.log10(m);
-      // Map dB range [-30, +30] to vertical axis
-      const y = height - ((db + 30) / 60) * height;
-      i === 0 ? ctx.moveTo(x, y) : ctx.lineTo(x, y);
-    });
-
-    ctx.strokeStyle = '#38acdd';
-    ctx.lineWidth = 2;
-    ctx.stroke();
-  };
-
-  return (
-    <View style={styles.container}>
-      {isLoading && <ActivityIndicator color="#fcfcff" />}
-
-      <Canvas
-        ref={canvasRef}
-        style={{ width: 600, height: 200, borderRadius: 6, marginTop: 16 }}
-      />
-      <Button onPress={handlePlayPause} title={isPlaying ? 'Pause' : 'Play'} color="#33488e" />
-
-      <ScrollView horizontal style={{ marginTop: 16 }}>
-        {FILTER_TYPES.map((type) => (
-          <Pressable
-            key={type}
-            onPress={() => setFilterType(type)}
-            style={[styles.filterButton, filterType === type && styles.filterButtonActive]}
-          >
-            <Text style={styles.labelText}>{type}</Text>
-          </Pressable>
-        ))}
-      </ScrollView>
-
-      <View style={styles.audioToggleContainer}>
-        <Text style={[styles.labelText, { marginRight: 8 }]}>Voice</Text>
-        <Pressable
-          onPress={() => setSelectedAudio(selectedAudio === 'audio1' ? 'audio2' : 'audio1')}
-          style={styles.toggleTrack}
-        >
-          <View
-            style={[
-              styles.toggleThumb,
-              { transform: [{ translateX: selectedAudio === 'audio1' ? 0 : 26 }] },
-            ]}
-          />
-        </Pressable>
-        <Text style={[styles.labelText, { marginLeft: 8 }]}>Music</Text>
-      </View>
-
-      <View style={{ marginTop: 16 }}>
-        <RangeSlider label="Frequency" value={filterFreq} min={10} max={5000} step={10} unit="Hz" onChange={setFilterFreq} />
-        {(filterType !== 'lowshelf' && filterType !== 'highshelf') && (
-          <RangeSlider label="Q" value={filterQ} min={0.1} max={20} step={0.1} onChange={setFilterQ} />
-        )}
-        {(filterType === 'peaking' || filterType === 'lowshelf' || filterType === 'highshelf') && (
-          <RangeSlider label="Gain" value={filterGain} min={-40} max={40} step={0.5} unit="dB" onChange={setFilterGain} />
-        )}
-      </View>
-    </View>
-  );
-};
-
-export default FrequencyResponseGraph;
-
-const styles = StyleSheet.create({
-  container: { flex: 1, padding: 16 },
-  canvas: { borderRadius: 6, marginTop: 16 },
-  filterButton: {
-    paddingVertical: 8,
-    paddingHorizontal: 12,
-    backgroundColor: '#6676aa',
-    borderRadius: 6,
-    marginRight: 8,
-  },
-  filterButtonActive: {
-    backgroundColor: '#33488e',
-  },
-  labelText: {
-    color: '#fcfcff',
-  },
-  audioToggleContainer: {
-    flexDirection: 'row',
-    alignItems: 'center',
-    marginTop: 16,
-  },
-  toggleTrack: {
-    width: 50,
-    height: 24,
-    borderRadius: 12,
-    backgroundColor: '#33488e',
-    padding: 2,
-    justifyContent: 'center',
-  },
-  toggleThumb: {
-    width: 20,
-    height: 20,
-    borderRadius: 10,
-    backgroundColor: '#6676aa',
-  },
-});
